requirements.txt
streamlit
openai
langchain
python-dotenv
PyPDF2
nltk
requests
pytest
chromadb
sentence-transformers
reportlab